Extract split_args() into utils.py that is imported to compiler and preprocessor On the fly import placeholders in handleStringArrayInitialisation to avoid circular import
