+++ ksp_compiler3/ksp_compiler.py
@@ -32,6 +32,7 @@ import time
 from preprocessor_plugins import pre_macro_functions, macro_iter_functions, substituteDefines, post_macro_iter_functions, post_macro_functions
 import json
 import copy
+import utils
 
 variable_prefixes = '$%@!?~'
 
@@ -115,35 +116,6 @@ def prefix_ID_with_ns(id, namespaces, function_parameter_names=None, force_prefi
         return ksp_ast.ID(id.lexinfo, identifier=prefix_with_ns(str(id), namespaces, function_parameter_names, force_prefixing))
     else:
         return id
-
-def split_args(arg_string, line):
-    '''converts eg. "x, y*(1+z), z" into a list ['x', 'y*(1+z)', 'z']'''
-    if arg_string.strip() == '':
-        return []
-    args = []
-    cur_arg = ''
-    unmatched_left_paren = 0
-    double_quote_on = False
-
-    for idx, ch in enumerate(arg_string + ','):    # extra ',' to include the last argument
-        # square brackets are also checked as there may be commas in them (for properties/2D arrays)
-        if ch == '\"' and (idx == 0 or arg_string[idx - 1] != '\\'):
-            double_quote_on = not double_quote_on
-        elif ch in ['(', '[']:
-            unmatched_left_paren += 1
-        elif ch in [')', ']']:
-            unmatched_left_paren -= 1
-        if ch == ',' and unmatched_left_paren == 0 and not double_quote_on:
-            cur_arg = cur_arg.strip()
-            if not cur_arg:
-                raise ParseException(line, 'Syntax error: empty argument in function call %s!' % arg_string)
-            args.append(cur_arg)
-            cur_arg = ''
-        else:
-            cur_arg += ch
-    if unmatched_left_paren:
-        raise ParseException(line, 'Error: unmatched parenthesis in function call %s!' % arg_string)
-    return args
 
 class ExceptionWithMessage(Exception):
     _message = None
@@ -504,7 +476,7 @@ def expand_macros(lines, macros, level=0, replace_raw=True):
 
                 macro = name2macro[macro_name]
                 if args:
-                    args = split_args(args[1:-1], line)
+                    args = utils.split_args(args[1:-1], line)
                 else:
                     args = []
